backend/requirements.txt
fastapi
uvicorn[standard]
chromadb
langchain
langchain-text-splitters
python-dotenv
ollama
openai
